raspberrypi/requirements.txt
Adafruit-Blinka==5.4.1
Adafruit-PlatformDetect==2.17.0
Adafruit-PureIO==1.1.6
attrs==20.2.0
Automat==20.2.0
bleak==0.8.0
colorzero==1.1
constantly==15.1.0
gpiozero==1.5.1
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
numpy==1.19.2
opencv-python==4.4.0.44
Pillow==7.2.0
pkg-resources==0.0.0
PyAudio==0.2.11
pyftdi==0.52.0
PyHamcrest==2.0.2
pyserial==3.4
pyusb==1.1.0
rpi-ws281x==4.2.4
RPi.GPIO==0.7.0
six==1.15.0
spidev==3.5
sysv-ipc==1.0.1
Twisted==20.3.0
txdbus==1.1.1
zope.interface==5.1.2
